Move Search functionality into a new class, update changelogs
