+++ packages/zowe-explorer-api/package.json
@@ -28,15 +28,15 @@
   },
   "dependencies": {
     "@types/vscode": "^1.53.2",
-    "@zowe/core-for-zowe-sdk": "^8.13.0",
-    "@zowe/imperative": "^8.13.0",
+    "@zowe/core-for-zowe-sdk": "^8.16.0",
+    "@zowe/imperative": "^8.16.0",
     "@zowe/secrets-for-zowe-sdk": "^8.10.4",
-    "@zowe/zos-console-for-zowe-sdk": "^8.13.0",
-    "@zowe/zos-files-for-zowe-sdk": "^8.13.0",
-    "@zowe/zos-jobs-for-zowe-sdk": "^8.13.0",
-    "@zowe/zos-tso-for-zowe-sdk": "^8.13.0",
-    "@zowe/zos-uss-for-zowe-sdk": "^8.13.0",
-    "@zowe/zosmf-for-zowe-sdk": "^8.13.0",
+    "@zowe/zos-console-for-zowe-sdk": "^8.16.0",
+    "@zowe/zos-files-for-zowe-sdk": "^8.16.0",
+    "@zowe/zos-jobs-for-zowe-sdk": "^8.16.0",
+    "@zowe/zos-tso-for-zowe-sdk": "^8.16.0",
+    "@zowe/zos-uss-for-zowe-sdk": "^8.16.0",
+    "@zowe/zosmf-for-zowe-sdk": "^8.16.0",
     "async-mutex": "^0.5.0",
     "deep-object-diff": "^1.1.9",
     "mustache": "^4.2.0",
