+++ packages/zowe-explorer-ftp-extension/package.json
@@ -52,9 +52,9 @@
     "vscode": "^1.79.0"
   },
   "dependencies": {
-    "@zowe/zos-files-for-zowe-sdk": "^8.13.0",
+    "@zowe/zos-files-for-zowe-sdk": "^8.16.0",
     "@zowe/zos-ftp-for-zowe-cli": "^3.0.0",
-    "@zowe/zos-jobs-for-zowe-sdk": "^8.13.0",
+    "@zowe/zos-jobs-for-zowe-sdk": "^8.16.0",
     "@zowe/zowe-explorer-api": "3.2.0-SNAPSHOT",
     "tmp": "0.2.3"
   },
