+++ packages/zowe-explorer/package.json
@@ -2016,11 +2016,11 @@
   },
   "dependencies": {
     "@vscode/codicons": "^0.0.35",
-    "@zowe/core-for-zowe-sdk": "^8.13.0",
+    "@zowe/core-for-zowe-sdk": "^8.16.0",
     "@zowe/secrets-for-zowe-sdk": "^8.10.4",
-    "@zowe/zos-files-for-zowe-sdk": "^8.13.0",
-    "@zowe/zos-jobs-for-zowe-sdk": "^8.13.0",
-    "@zowe/zosmf-for-zowe-sdk": "^8.13.0",
+    "@zowe/zos-files-for-zowe-sdk": "^8.16.0",
+    "@zowe/zos-jobs-for-zowe-sdk": "^8.16.0",
+    "@zowe/zosmf-for-zowe-sdk": "^8.16.0",
     "@zowe/zowe-explorer-api": "3.2.0-SNAPSHOT",
     "dayjs": "^1.11.10",
     "fs-extra": "8.0.1",
