+++ packages/zowe-explorer/src/trees/dataset/DatasetActions.ts
@@ -12,19 +12,7 @@
 import * as vscode from "vscode";
 import * as zosfiles from "@zowe/zos-files-for-zowe-sdk";
 import * as path from "path";
-import {
-    Gui,
-    imperative,
-    IZoweDatasetTreeNode,
-    Validation,
-    Table,
-    TableBuilder,
-    TableViewProvider,
-    Types,
-    FsAbstractUtils,
-    ZoweScheme,
-    ZoweExplorerApiType,
-} from "@zowe/zowe-explorer-api";
+import { Gui, imperative, IZoweDatasetTreeNode, Validation, Types, FsAbstractUtils, ZoweScheme, ZoweExplorerApiType } from "@zowe/zowe-explorer-api";
 import { ZoweDatasetNode } from "./ZoweDatasetNode";
 import { DatasetUtils } from "./DatasetUtils";
 import { DatasetFSProvider } from "./DatasetFSProvider";
@@ -40,13 +28,6 @@ import { SharedUtils } from "../shared/SharedUtils";
 import { FilterItem } from "../../management/FilterManagement";
 import { AuthUtils } from "../../utils/AuthUtils";
 import { Definitions } from "../../configuration/Definitions";
-
-interface ISearchOptions {
-    node: IZoweDatasetTreeNode;
-    pattern: string;
-    searchString: string;
-    caseSensitive: boolean;
-}
 
 export class DatasetActions {
     public static typeEnum: zosfiles.CreateDataSetTypeEnum;
@@ -1875,272 +1856,5 @@ export class DatasetActions {
         } else if (SharedContext.isDs(node) || SharedContext.isPds(node) || SharedContext.isMigrated(node) || SharedContext.isVsam(node)) {
             await vscode.env.clipboard.writeText(node.label as string);
         }
-    }
-
-    public static async search(context: vscode.ExtensionContext, node: IZoweDatasetTreeNode): Promise<void> {
-        const isSessionNotFav = SharedContext.isSessionNotFav(node);
-        const generateFullUri = (isSessionNotFav && node.pattern != null) || SharedContext.isFavoriteSearch(node);
-        const pattern = isSessionNotFav ? node.pattern : (node.label as string);
-
-        // There may not be a pattern on a session node if there is no filter applied. Warn if this is the case.
-        if (!pattern) {
-            Gui.errorMessage(vscode.l10n.t("No search pattern applied. Search for a pattern and try again."));
-            return;
-        }
-
-        // Figure out what text we are looking for.
-        const searchString = await Gui.showInputBox({ prompt: vscode.l10n.t("Enter the text to search for.") });
-        if (!searchString) {
-            return;
-        }
-
-        // Handle various search options
-        const caseSensitiveQuickPick = await Gui.showQuickPick(
-            [
-                {
-                    label: vscode.l10n.t("Case Sensitive"),
-                    description: vscode.l10n.t("Perform the search with case sensitivity"),
-                    iconPath: new vscode.ThemeIcon("case-sensitive"),
-                },
-            ],
-            {
-                title: vscode.l10n.t("Search Options"),
-                placeHolder: vscode.l10n.t("Select search options"),
-                ignoreFocusOut: true,
-                canPickMany: true,
-            }
-        );
-
-        // Case sensitivity
-        if (caseSensitiveQuickPick == undefined) {
-            return;
-        }
-        const caseSensitive = caseSensitiveQuickPick.some((option) => option.label === vscode.l10n.t("Case Sensitive"));
-
-        // Perform the actual search.
-        const response: zosfiles.IZosFilesResponse = await Gui.withProgress(
-            {
-                location: vscode.ProgressLocation.Notification,
-                title: vscode.l10n.t({ message: 'Searching for "{0}"', args: [searchString], comment: "The string to search data sets for." }),
-                cancellable: true,
-            },
-            async (progress, token) => {
-                return this.performSearch(progress, token, { node, pattern, searchString, caseSensitive });
-            }
-        );
-
-        // Either the user cancelled the search, or a catastrophic error occurred and error handling has already been done.
-        if (response === undefined) {
-            return;
-        }
-
-        // Prepare a list of matches in a format the table expects
-        const matches = this.getSearchMatches(node, response, generateFullUri, searchString);
-
-        // Prepare the table for display to the user
-        const table = new TableBuilder(context)
-            .title(vscode.l10n.t({ message: 'Search Results for "{0}"', args: [searchString], comment: "The string to search data sets for." }))
-            .options({
-                autoSizeStrategy: { type: "fitCellContents" },
-                pagination: true,
-                rowSelection: "multiple",
-                selectEverything: true,
-                suppressRowClickSelection: true,
-            })
-            .isView()
-            .rows(...matches)
-            .columns(
-                ...[
-                    {
-                        field: "name",
-                        headerName: vscode.l10n.t("Data Set Name"),
-                        filter: true,
-                        initialSort: "asc",
-                    } as Table.ColumnOpts,
-                    {
-                        field: "position",
-                        headerName: vscode.l10n.t("Position"),
-                        filter: false,
-                    },
-                    {
-                        field: "contents",
-                        headerName: vscode.l10n.t("Contents"),
-                        filter: true,
-                    },
-                    {
-                        field: "actions",
-                        hide: true,
-                    },
-                ]
-            )
-            .addRowAction("all", {
-                title: vscode.l10n.t("Open"),
-                command: "open",
-                callback: {
-                    fn: DatasetActions.openSearchAtLocation,
-                    typ: "multi-row",
-                },
-                type: "secondary",
-            })
-            .build();
-
-        // Show the table to the user.
-        await TableViewProvider.getInstance().setTableView(table);
-    }
-
-    private static async openSearchAtLocation(this: void, _view: Table.View, data: Record<number, Table.RowData>): Promise<void> {
-        const childrenToOpen = Object.values(data);
-        if (childrenToOpen.length === 0) {
-            return;
-        }
-
-        for (const child of childrenToOpen) {
-            // Get the URI, look it up so the Filesystem provider is aware of it, then display the document to the user
-            const childUri = vscode.Uri.from({ scheme: ZoweScheme.DS, path: child.uri as string });
-            await DatasetFSProvider.instance.remoteLookupForResource(childUri);
-            void Gui.showTextDocument(childUri, { preview: false }).then(
-                (editor) => {
-                    // Highlight the searched for text so the user can easily find it
-                    const startPosition = new vscode.Position((child.line as number) - 1, (child.column as number) - 1);
-                    const endPosition = new vscode.Position(
-                        (child.line as number) - 1,
-                        (child.column as number) - 1 + (child.searchString as string).length
-                    );
-                    editor.selection = new vscode.Selection(startPosition, endPosition);
-                },
-                (err) => {
-                    Gui.errorMessage(err.message);
-                }
-            );
-        }
-    }
-
-    private static async continueSearchPrompt(this: void, dataSets: zosfiles.IDataSet[]): Promise<boolean> {
-        const MAX_DATASETS = 50;
-
-        // If there are 50 matches or under, do not prompt the user for confirmation to continue
-        if (dataSets.length <= MAX_DATASETS) {
-            return true;
-        }
-
-        // There are 51 or more results, which may take a long time. Make sure the user is prepared for that.
-        const resp = await Gui.infoMessage(
-            vscode.l10n.t({
-                message: "Are you sure you want to search {0} data sets and members?",
-                args: [dataSets.length.toString()],
-                comment: "The number of data sets that are about to be searched",
-            }),
-            {
-                items: [vscode.l10n.t("Continue")],
-                vsCodeOpts: { modal: true },
-            }
-        );
-
-        return resp === vscode.l10n.t("Continue");
-    }
-
-    private static async performSearch(progress: any, token: vscode.CancellationToken, options: ISearchOptions): Promise<zosfiles.ISearchResponse> {
-        const profile = options.node.getProfile();
-        const mvsApi = ZoweExplorerApiRegister.getMvsApi(profile);
-        let response: zosfiles.ISearchResponse;
-        if (token.isCancellationRequested) {
-            Gui.showMessage(DatasetActions.localizedStrings.opCancelled);
-            return;
-        }
-
-        // Prepare a hacky task object that updates the progress bar on the GUI.
-        // Pass that into the MVS API call.
-        let realPercentComplete = 0;
-        let realTotalEntries = 100;
-        const task: imperative.ITaskWithStatus = {
-            set percentComplete(value: number) {
-                realPercentComplete = value;
-                // eslint-disable-next-line no-magic-numbers
-                Gui.reportProgress(progress, realTotalEntries, Math.floor((value * realTotalEntries) / 100), "");
-            },
-            get percentComplete(): number {
-                return realPercentComplete;
-            },
-            statusMessage: "",
-            stageName: 0, // TaskStage.IN_PROGRESS - https://github.com/kulshekhar/ts-jest/issues/281
-        };
-
-        try {
-            // Perform the actual search
-            response = await mvsApi.searchDataSets({
-                pattern: options.pattern,
-                searchString: options.searchString,
-                progressTask: task,
-                mainframeSearch: false,
-                caseSensitive: options.caseSensitive,
-                continueSearch: function intercept(dataSets: zosfiles.IDataSet[]) {
-                    realTotalEntries = dataSets.length;
-                    return DatasetActions.continueSearchPrompt(dataSets);
-                },
-                abortSearch: function abort() {
-                    return token.isCancellationRequested;
-                },
-            });
-
-            // The user cancelled the search
-            if (response.success === false && response.commandResponse?.includes("cancelled")) {
-                return;
-            }
-
-            // If there is no API response and success is false, the search didn't even begin searching data sets, and stopped during listing.
-            // Return an error to the user since we have no useful data. Otherwise, display partial results.
-            if (response.success === false) {
-                ZoweLogger.error(response.errorMessage);
-                Gui.errorMessage(response.errorMessage);
-            }
-            return response.apiResponse != null ? response : undefined;
-        } catch (err) {
-            // Something catastrophic happened that was not handled (i.e. bad credentials).
-            await AuthUtils.errorHandling(err);
-            return;
-        }
-    }
-
-    private static getSearchMatches(
-        node: IZoweDatasetTreeNode,
-        response: any,
-        generateFullUri: boolean,
-        searchString: string
-    ): Record<string, any>[] {
-        const matches = response.apiResponse;
-        const newMatches: object[] = [];
-
-        // Take in the API response, and iterate through the list of matched data sets and members
-        for (const ds of matches) {
-            const dsn = ds.dsn as string;
-            const member = ds.member as string;
-            const extension = DatasetUtils.getExtension(ds.dsn);
-
-            let name: string = dsn;
-            let uri = generateFullUri ? path.posix.join(node.getSessionNode().resourceUri.path, dsn) : node.resourceUri.path;
-
-            if (member) {
-                uri = uri + "/" + member;
-                name = name + "(" + member + ")";
-            }
-
-            if (extension != null) {
-                uri += extension;
-            }
-
-            // The data set or member might have multiple matches itself. Display each one as a separate item.
-            for (const match of ds.matchList) {
-                newMatches.push({
-                    name,
-                    line: match.line as number,
-                    column: match.column as number,
-                    position: (match.line as number).toString() + ":" + (match.column as number).toString(),
-                    contents: match.contents,
-                    uri,
-                    searchString,
-                });
-            }
-        }
-        return newMatches;
     }
 }
